docs: modified docs according to my review
- demo/angularjs/components/button/doc.md
"allow" is almost never necessary. Reworked descriptions to remove it.
Harmonized descriptions of modes.
Clarified the Low emphasis description.

- demo/angularjs/components/checkbox/doc.md
Made an inline list for the enumeration (none, one or multiple)

- demo/angularjs/components/chip/doc.md
Harmonized formulation.

- demo/angularjs/components/dialog/doc.md
Split the first sentence and specified title and buttons as default.
Harmonized the descriptions

- demo/angularjs/components/dropdown/doc.md
Removed descriptions and modified titles to be more specific.

- demo/angularjs/components/list/doc.md
Simplified sentence

- demo/angularjs/components/progress-tracker/doc.md
moved the helper text mention to the main description
simplified the states description
clairified the error description

- demo/angularjs/components/radio-button/doc.md
simplified description

- demo/angularjs/components/select/doc.md
Avoided allow as it is not specific enough

- demo/angularjs/components/switch/doc.md
Simplified description

- demo/angularjs/components/table/doc.md
Simplified description, removed example as it should be in the demo

- demo/angularjs/components/text-field/doc.md
Avoid "let", simplified the description

- demo/angularjs/components/toolbar/doc.md
Simplified the description
Clarified the mention of three sections by describing the three sections

- demo/angularjs/components/user-block/doc.md
Aligned the description of sizes on the other desc of sizes
Typed Vertical and Extended as layouts
Clarified Extended layout.